predict_requirements.txt
torch
torchvision
tqdm
piexif
Pillow
matplotlib
configargparse
numpy
exifread
rawpy
opencv-python
